requirements.txt
python-dotenv
google-generativeai
requests
slack_sdk
Flask
beautifulsoup4==4.12.3
lxml==5.2.2
scikit-learn
pandas